requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.5.2
aspose-words==22.10.0
aspose-words-cloud==22.10.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
bytesbufio==1.0.3
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
cobble==0.1.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
comtypes==1.1.14
cryptography==38.0.1
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
deprecation==2.1.0
distlib==0.3.5
distro==1.7.0
distro-info===1.1build1
Django==4.0
dnspython==2.1.0
docx2pdf==0.1.8
docx2txt==0.8
duplicity==0.8.21
emoji==2.1.0
entrypoints==0.4
fasteners==0.14.1
filelock==3.7.1
fire==0.4.0
fonttools==4.34.4
frozenlist==1.3.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
glob2==0.7
gpg===1.16.0-unknown
gTTS==2.2.4
gyp==0.1
httplib2==0.20.2
idna==3.3
img2pdf==0.4.4
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
mammoth==1.5.0
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
mysql-connector-python==8.0.29
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandoc==2.2
paramiko==2.9.3
pdf2docx==0.5.5
pdf2image==1.16.0
pdfminer.six==20220524
pdfplumber==0.7.5
pexpect==4.8.0
pikepdf==6.2.1
Pillow==9.2.0
platformdirs==2.5.2
plumbum==1.8.0
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
pyarrow==10.0.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydeck==0.8.0
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyMuPDF==1.20.1
PyNaCl==1.5.0
pypandoc==1.9
pyparsing==2.4.7
PyPDF2==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.19.2
pyTelegramBotAPI==4.7.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.10.1
rich==12.6.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
ssh-import-id==5.11
streamlit==1.16.0
systemd-python==234
tabula-py==2.4.0
telegram-menu==0.3.15
termcolor==1.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.16.3
wadllib==1.3.6
Wand==0.6.10
watchdog==2.2.0
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
